+++ react/src/components/AtomForm/index.js
@@ -1,9 +1,10 @@
 import React from 'react'
-import { Button, Input } from 'reactstrap'
+import { Input } from 'reactstrap'
 import { isEqual } from 'lodash'
 import ReactModal from 'react-modal'
 
 import NestedModelControls from 'components/NestedModelControls'
+import FolioConsoleUiButtons from 'components/FolioConsoleUiButtons'
 import FolioConsoleUiButton from 'components/FolioConsoleUiButton'
 import FolioUiIcon from 'components/FolioUiIcon'
 
@@ -269,7 +270,25 @@ class AtomForm extends React.PureComponent {
               } else {
                 zIndex -= 1
 
-                if (input === 'ASSOCIATIONS') {
+                if (input === 'aside_attachments') {
+                  return (
+                    <AtomFormCardColumns data-debug='columns' style={makeStyle()}>
+                      <AtomFormCardColumn data-debug='column' dontGrow>
+                        {fillOutput('ATTACHMENTS')}
+                      </AtomFormCardColumn>
+                      <AtomFormCardColumn data-debug='column'>
+                        <AtomFormCardRows data-debug='rows' style={makeStyle()}>
+                          <AtomFormCardRow data-debug='row' style={makeStyle()}>
+                            {fillOutput('ASSOCIATIONS')}
+                          </AtomFormCardRow>
+                          <AtomFormCardRow data-debug='row' style={makeStyle()}>
+                            {fillOutput('STRUCTURE')}
+                          </AtomFormCardRow>
+                        </AtomFormCardRows>
+                      </AtomFormCardColumn>
+                    </AtomFormCardColumns>
+                  )
+                } else if (input === 'ASSOCIATIONS') {
                   return (
                     <Associations
                       atom={atom}
@@ -372,12 +391,17 @@ class AtomForm extends React.PureComponent {
             )
           })}
 
-          {addButtons.map((type) => (
-            <Button color='success' type='button' className='mr-2' onClick={() => { this.props.addAtom(type.type) }} key={type.type}>
-              <FolioUiIcon name='plus' />
-              {type.title}
-            </Button>
-          ))}
+          <FolioConsoleUiButtons>
+            {addButtons.map((type) => (
+              <FolioConsoleUiButton
+                variant='success'
+                onClick={() => { this.props.addAtom(type.type) }}
+                key={type.type}
+                label={type.title}
+                icon='plus'
+              />
+            ))}
+          </FolioConsoleUiButtons>
         </div>
 
         <ReactModal