+++ test/dummy/app/components/dummy/ui/image_component.rb
@@ -1,7 +1,9 @@
 # frozen_string_literal: true
 
 class Dummy::Ui::ImageComponent < ApplicationComponent
-  bem_class_name :lightbox, :contain, :cover, :hover_zoom, :round
+  include Folio::CellLightbox
+
+  bem_class_name :lightbox, :contain, :cover, :hover_zoom, :round, :border_radius
 
   RETINA_MULTIPLIER = 2
 
@@ -18,8 +20,10 @@ class Dummy::Ui::ImageComponent < ApplicationComponent
                  style: nil,
                  spacer_background: true,
                  round: false,
+                 border_radius: false,
                  additional_html: nil,
                  class_name: nil)
+    @placement = placement
     @size = size
     @lightbox = lightbox
     @contain = contain
@@ -33,6 +37,7 @@ class Dummy::Ui::ImageComponent < ApplicationComponent
     @additional_html = additional_html
     @class_name = class_name
     @round = round
+    @border_radius = border_radius
     @lazy = lazy
     @data = set_data(placement)
   end
@@ -200,34 +205,20 @@ class Dummy::Ui::ImageComponent < ApplicationComponent
       tag: :div,
       class: class_names,
       style: wrap_style,
-      "data-width" => thumbnail_width,
-      "data-height" => thumbnail_height,
+      data: {
+        width: thumbnail_width,
+        height: thumbnail_height,
+      }
     }
 
     if @href
       h[:tag] = :a
       h[:href] = @href
-      # TODO lightbox
-      # elsif model && lightboxable?
-      #   if @lightbox.is_a?(Hash)
-      #     h = h.merge(@lightbox)
-      #   elsif model.is_a?(Folio::FilePlacement::Base)
-      #     h = h.merge(lightbox(model))
-      #     h["data-lightbox-caption"] = @title if @title
-      #   else
-      #     h = h.merge(lightbox_from_image(model))
-      #   end
+    elsif @lightbox
+      h[:data].merge!(stimulus_lightbox_item(@placement, title: @title))
     end
 
     h
-  end
-
-  def self.fixed_height_mobile_ratio
-    0.862
-  end
-
-  def self.fixed_height_default_width
-    160
   end
 
   def thumbnail_size