+++ test/dummy/app/controllers/dummy/ui_controller.rb
@@ -115,6 +115,7 @@ class Dummy::UiController < ApplicationController
         add_breadcrumb "UI", dummy_ui_path
 
         if action_name != "show"
+          add_breadcrumb "UI", dummy_ui_path
           add_breadcrumb action_name.capitalize, send("#{action_name}_dummy_ui_path")
         end