+++ test/dummy/db/seeds.rb
@@ -33,37 +33,6 @@ destroy_all Folio::Site
 destroy_all Dummy::Blog::Article
 destroy_all Dummy::Blog::Topic
 
-def unsplash_pic(square = false)
-  puts "Creating unsplash pic"
-
-  image = Folio::File::Image.new(site: ::Folio.main_site)
-  scale = 0.5 + rand / 2
-  w = (scale * 2560).to_i
-  h = (square ? scale * 2560 : scale * 1440).to_i
-  image.file_url = "https://picsum.photos/#{w}/#{h}/?random"
-  image.tag_list = "unsplash, seed"
-  image.file_name = "unsplash.jpg"
-  image.file_width = w
-  image.file_height = h
-  image.save!
-
-  puts "Created unsplash pic"
-
-  image
-end
-
-def file_pic(file_instance)
-  puts "Creating file pic"
-
-  image = Folio::File::Image.new(site: ::Folio.main_site)
-  image.file = file_instance
-  image.save!
-
-  puts "Created file pic"
-
-  image
-end
-
 puts "Creating Folio::Site"
 Folio::Site.create!(title: "Sinfin.digital",
                     domain: "sinfin.localhost",
@@ -84,72 +53,29 @@ Folio::Site.create!(title: "Sinfin.digital",
                     })
 puts "Created Folio::Site"
 
-2.times { unsplash_pic }
-
-puts "Creating about page"
-about = Folio::Page.create!(title: "O nás",
-                            published: true,
-                            published_at: 1.month.ago,
-                            site: ::Folio.main_site)
-about.cover = Folio::File::Image.first
-about.image_placements.each { |ip|
-  name = "Lorem Ipsum"
-  ip.update_attributes!(alt: name, title: "Portrait of #{name}")
-}
-puts "Created about page"
-
-puts "Creating more pages"
-night_sky = Folio::Page.create!(title: "Noční obloha",
-                                published: true,
-                                published_at: 1.month.ago,
-                                locale: :cs,
-                                site: ::Folio.main_site)
-night_photo = File.new(Folio::Engine.root.join("test/fixtures/folio/photos/night.jpg"))
-night_sky.cover = file_pic(night_photo)
-1.times { night_sky.images << file_pic(night_photo) }
-
-reference = Folio::Page.create!(title: "Reference",
-                                published: true,
-                                published_at: 1.day.ago,
-                                site: ::Folio.main_site)
-Folio::Page.create!(title: "Smart Cities",
-                    published: true,
-                    published_at: 1.month.ago,
-                    site: ::Folio.main_site)
-vyvolejto = Folio::Page.create!(title: "Vyvolej.to",
-                                published: true,
-                                published_at: 1.month.ago,
-                                site: ::Folio.main_site)
-iptc_test = File.new(Folio::Engine.root.join("test/fixtures/folio/photos/downsized-exif-samples/jpg/tests/46_UnicodeEncodeError.jpg"))
-vyvolejto.cover = file_pic(iptc_test)
-
-Folio::Page.create!(title: "Hidden", published: false, site: ::Folio.main_site)
-Folio::Page.create!(title: "DAM", published: true, site: ::Folio.main_site)
-puts "Created more pages"
-
 puts "Creating Dummy::Menu::Nestable"
 menu = Dummy::Menu::Nestable.create!(locale: :cs, title: "Nestable", site: ::Folio.main_site)
 
 root = Folio::MenuItem.create!(menu:,
                                title: "Reference",
-                               target: reference,
+                               url: "./",
                                position: 0)
 
 child = Folio::MenuItem.create!(menu:,
                                 title: "Podreference",
-                                target: reference,
+                                url: "./",
                                 position: 1,
                                 parent: root)
 
 Folio::MenuItem.create!(menu:,
                         title: "Podreference",
-                        target: reference,
+                        url: "./",
                         position: 2,
                         parent: child)
 
 Folio::MenuItem.create!(menu:,
                         title: "About",
-                        target: about,
+                        url: "./",
                         position: 3)
 puts "Created Dummy::Menu::Nestable"
 
@@ -158,18 +84,18 @@ menu = Dummy::Menu::Stylable.create!(locale: :cs, title: "Stylable", site: ::Fol
 
 Folio::MenuItem.create!(menu:,
                         title: "Reference",
-                        target: reference,
+                        url: "./",
                         position: 0)
 
 Folio::MenuItem.create!(menu:,
                         title: "About red",
-                        target: about,
+                        url: "./",
                         position: 1,
                         style: "red")
 
 Folio::MenuItem.create!(menu:,
                         title: "About",
-                        target: about,
+                        url: "./",
                         position: 1)
 puts "Created Dummy::Menu::Stylable"
 
@@ -257,48 +183,3 @@ menu = Dummy::Menu::Footer.create!(locale: :cs, title: "Footer", site: ::Folio.m
 end
 
 puts "Created Dummy::Menu::Footer"
-
-images = Folio::File::Image.by_site(::Folio.main_site).tagged_with("unsplash").to_a
-
-puts "Creating Dummy::Blog::Topic"
-
-topics = 5.times.map do
-  topic = Dummy::Blog::Topic.create!(locale: "cs",
-                                     published: true,
-                                     title: Faker::Hipster.words(number: rand(1..3)).join(" ").capitalize,
-                                     cover: images.sample)
-
-  print(".")
-
-  topic
-end
-
-puts "\nCreated Dummy::Blog::Topic"
-
-puts "Creating Dummy::Blog::Article"
-
-20.times do |i|
-  Dummy::Blog::Article.create!(locale: "cs",
-                               published: true,
-                               title: Faker::Hipster.words(number: rand(1..3)).join(" ").capitalize,
-                               perex: Faker::Hipster.sentence,
-                               cover: images.sample,
-                               topics: topics.sample(rand(1..topics.size)),
-                               published_at: Time.zone.now - i.days - i.hours - i.minutes)
-  print(".")
-end
-
-puts "\nCreated Dummy::Blog::Article"
-
-if Rails.env.development?
-  puts "Creating test@test.test superadmin user"
-
-  Folio::User.create!(email: "test@test.test",
-                      password: "test@test.test",
-                      superadmin: true,
-                      first_name: "Test",
-                      last_name: "Dummy",
-                      confirmed_at: Time.current)
-
-  puts "Created test@test.test superadmin user"
-end