+++ test/dummy/test/components/dummy/atom/title_component_test.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+require "test_helper"
+
+class Dummy::Atom::TitleComponentTest < Folio::ComponentTest
+  def test_render
+    atom = create_atom(Dummy::Atom::Title, :title)
+
+    render_inline(Dummy::Atom::TitleComponent.new(atom:))
+
+    assert_selector(".d-atom-title")
+  end
+end