+++ test/dummy/test/components/dummy/molecule/cards/padded_image_component_test.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+require "test_helper"
+
+class Dummy::Molecule::Cards::PaddedImageComponentTest < Folio::ComponentTest
+  def test_render
+    atoms = [create_atom(Dummy::Atom::Cards::PaddedImage, :title, :url, :cover)]
+
+    render_inline(Dummy::Molecule::Cards::PaddedImageComponent.new(atoms:))
+
+    assert_selector(".d-molecule-cards-padded-image")
+  end
+end