+++ test/dummy/test/components/dummy/molecule/cards/small_component_test.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+require "test_helper"
+
+class Dummy::Molecule::Cards::SmallComponentTest < Folio::ComponentTest
+  def test_render
+    atoms = [create_atom(Dummy::Atom::Cards::Small)]
+
+    render_inline(Dummy::Molecule::Cards::SmallComponent.new(atoms:))
+
+    assert_selector(".d-molecule-cards-small")
+  end
+end